Move test suite to correct module
